app/requirements.txt
amqp==2.3.2
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.4
boto3==1.9.33
botocore==1.12.33
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-celery-beat==1.2.0
django-rest-framework==0.1.0
django-timezone-field==3.0
djangorestframework==3.9.0
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
kombu==4.2.1
more-itertools==4.3.0
mypy==0.641
mypy-extensions==0.4.1
oauthlib==2.1.0
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
PyJWT==1.6.4
pytest==3.9.3
pytest-django==3.4.3
python-crontab==2.3.5
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.6
redis==2.10.6
requests==2.20.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
typed-ast==1.1.0
urllib3==1.24
uWSGI==2.0.17.1
vine==1.1.4
